integrations/discord_bot/requirements.txt
discord.py==2.4.0
python-dotenv==1.0.0
memu-py==1.4.0
requests>=2.0.0
